requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
openai>=1.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2>=3.1.0
PyPDF2==3.0.1
python-magic==0.4.27
reportlab>=4.1.0 